Extract FpsCounter
